+++ packages/layer-composer/src/generators/types.ts
@@ -2,6 +2,8 @@ import { FeatureCollection } from 'geojson'
 import { StringUnitLength } from 'luxon'
 import { AnySourceData, Layer } from '@globalfishingwatch/mapbox-gl'
 import { Segment } from '@globalfishingwatch/data-transforms'
+import { AggregationOperation } from '@globalfishingwatch/fourwings-aggregate'
+import { IntervalOption } from './heatmap/util/time-chunks'
 
 export enum Type {
   Background = 'BACKGROUND',
@@ -38,6 +40,7 @@ export type AnyData = FeatureCollection | Segment[] | RawEvent[] | Ruler[] | nul
 export interface GeneratorLegend {
   label?: string
   unit?: string
+  color?: string
 }
 
 export interface GeneratorMetadata {
@@ -138,6 +141,10 @@ export type TileClusterEventType = 'encounter' | 'loitering' | 'port'
  */
 export interface TileClusterGeneratorConfig extends GeneratorConfig {
   type: Type.TileCluster
+  /**
+   * Defines the maximum zoom that returns clusters
+   */
+  maxZoomCluster?: number
   /**
    * Sets the color of the line https://docs.mapbox.com/mapbox-gl-js/style-spec/layers/#paint-fill-fill-color
    */
@@ -267,6 +274,9 @@ export interface HeatmapAnimatedGeneratorConfig extends GeneratorConfig {
   interactive?: boolean
   staticStart?: string
   staticEnd?: string
+  interval?: IntervalOption
+  aggregationOperation?: AggregationOperation
+  breaksMultiplier?: number
 }
 
 export type AnyGeneratorConfig =
@@ -336,6 +346,8 @@ export interface HeatmapAnimatedGeneratorSublayer {
   filter?: string
   colorRamp: ColorRampsIds
   visible?: boolean
+  breaks?: number[]
+  legend?: GeneratorLegend
 }
 
 // ---- Heatmap Generator color ramps types
@@ -363,4 +375,6 @@ export enum HeatmapAnimatedMode {
   Blob = 'blob',
   // Represents value in 3D stacked bars. Works with 0 - n sublayers
   Extruded = 'extruded',
+  // Just show raw value ffor 1 sublayer
+  Single = 'single',
 }